Move cytoscape api
