+++ src/models/graphSelection.ts
@@ -1,5 +1,5 @@
 import { Quad } from 'n3';
-import { RdfNodeDataDefinition } from './cytoscapeApi.types';
+import { RdfNodeDataDefinition } from '../cytoscape-api/cytoscapeApi.types';
 
 export class GraphSelection {
 	constructor(readonly individuals: RdfNodeDataDefinition[], readonly edges: Quad[]) {}
